+++ js/data.js
@@ -1,9 +1,9 @@
-//модуль, который создаёт данные
+// модуль для генерации временных данных для разработки
+import { getRandomArrEl, getRandomNum, getRandomInt, getRandomRepeatArrEl, idComment } from './util-uniq-random.js';
 
-// id для фото - Идентификаторы не должны повторяться
+
 const idPhoto = [1, 2, 3, 4, 5, 6, 7, 8, 9, 10, 11, 12, 13, 14, 15, 16, 17, 18, 19, 20, 21, 22, 23, 24, 25];
 
-//адрес картинки - Адреса картинок не должны повторяться.
 const urlPhoto = [
   'photos/1.jpg',
   'photos/2.jpg',
@@ -32,7 +32,6 @@ const urlPhoto = [
   'photos/25.jpg',
 ];
 
-// описание фотографии
 const descriptionPhoto = [
   'Это рай на земле',
   'Наконец-то отпуск',
@@ -61,11 +60,6 @@ const descriptionPhoto = [
   'Жизнь бывает и такой',
 ];
 
-//количество лайков - случайное число от 15 до 200
-
-/*------------------*/
-// Комментарии
-// текст комментария - одно или два случайных предложения
 const messageComment = [
   'Всё отлично!',
   'В целом всё неплохо. Но не всё.',
@@ -75,7 +69,6 @@ const messageComment = [
   'Лица у людей на фотке перекошены, как будто их избивают. Как можно было поймать такой неудачный момент?!',
 ];
 
-// имя комментатора - случайное
 const nameComment = [
   'Петя',
   'Ваня',
@@ -89,4 +82,28 @@ const nameComment = [
   'Ильюша',
 ];
 
-export {idPhoto, urlPhoto, descriptionPhoto, messageComment, nameComment};
+const photoCount = 25;
+
+const createComment = () =>
+  ({
+    idComment: getRandomArrEl(idComment),
+    avatar: `img/avatar-${getRandomNum(1, 6)}.svg`,
+    message: messageComment.slice(0, (getRandomInt(1, 6))).join(', '),
+    name: getRandomRepeatArrEl(nameComment),
+  });
+
+
+const createPhoto = () =>
+  ({
+    id: getRandomArrEl(idPhoto),
+    url: getRandomArrEl(urlPhoto),
+    description: getRandomArrEl(descriptionPhoto),
+    likes: getRandomNum(15, 200),
+    comments: createComment(),
+  });
+
+
+const photoObjects = new Array(photoCount).fill(null).map(() => createPhoto());
+photoObjects();
+
+// export {photoObjects};
