+++ js/main.js
@@ -1,4 +1,2 @@
 import './util-functions/util.js';
-import './util-functions/util-object-array.js'
 import './util-functions/util-string.js'
-
